requirements.txt
streamlit
langchain
openai
python-dotenv
sentence-transformers
chromadb
pymysql
mysql-connector-python
